Move error handling to `HTTPErrorHandler`
