+++ domains/core/testhelpers/helpers.go
@@ -4,7 +4,10 @@ Package testhelpers contains various test helper functions
 package testhelpers
 
 import (
+	"bytes"
 	"crypto/rand"
+	"net/http/httptest"
+	"testing"
 )
 
 const alphanum = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
@@ -20,4 +23,22 @@ func RandomString(prefix string, length int) string {
 	}
 
 	return prefix + string(bytes)
+}
+
+// ClosingRecorder creates new httptest.ResponseRecorder and closing body at cleanup.
+func ClosingRecorder(t *testing.T) *httptest.ResponseRecorder {
+	t.Helper()
+
+	recorder := &httptest.ResponseRecorder{Body: new(bytes.Buffer)}
+
+	t.Cleanup(func() {
+		result := recorder.Result()
+		if result == nil {
+			return
+		}
+
+		_ = result.Body.Close()
+	})
+
+	return recorder
 }
